Componentizando class navigation
