+++ pubspec.yaml
@@ -52,6 +52,8 @@ dev_dependencies:
 
 # The following section is specific to Flutter packages.
 flutter:
+  assets:
+    - assets/images/Logo.png
 
   # The following line ensures that the Material Icons font is
   # included with your application, so that you can use the icons in
